Test that every module can be parsed
We can use this as a check to see just how well we've done with corefn.
If we fail here, someone consuming the library would also fail.
This should keep us honest about how much of the ecosystem we support.

This is currently failing on `CoreFn.Names`,
so we're already getting mileage out of this check.