Make travis build the tests as well
Since we're reading everything in the `output` directory,
tests have to have their corefn produced as well.
Normally, `pulp` doesn't build the `test` directory.
So, we tell it to include that one in the build as well.
Then, when we run `pulp test`, it ought to work for every compiled module.